requirements.txt
click==8.1.2
dnspython==2.2.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
loguru==0.6.0
MarkupSafe==2.1.1
PyJWT==1.4.2
pymongo==4.0.2
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.1.1
